su: Add back mistakenly removed code
Commit 71839266b7a3b18bd0824e1896c8dab7ed217508 removed this part
of the code because it was mistakenly thought as a samsung hack.

This was a misunderstanding because of the highly outdated comment,
which was referring to commit 62843e15043a07952ec2d16e50db2bbefeebb571
hacks which were later removed in 379371bcff3f18b4a16beb37a35c2b211a941781

This should fix running su when you are already root.

Change-Id: Id548cb8499eaab943c18a5ed564da8f61e4e390c
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>